2.2.3 Moved IntArrayStack
